# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "這是後端 API 文件",
        "title": "interchange API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-User"
                ],
                "summary": "取得用戶列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "頁碼",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每頁筆數",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "角色",
                        "name": "role",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "狀態",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/interchange_internal_dto.UserResponseDto"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-User"
                ],
                "summary": "新增用戶",
                "parameters": [
                    {
                        "description": "用戶資訊",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_dto.CreateUserDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_dto.UserResponseDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/users/{userID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-User"
                ],
                "summary": "取得單一用戶資訊",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_dto.UserResponseDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-User"
                ],
                "summary": "更新用戶",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "用戶更新資訊",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_dto.UpdateUserDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-User"
                ],
                "summary": "刪除用戶",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/admin/users/{userID}/role": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-User"
                ],
                "summary": "更新用戶角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "角色資訊",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_dto.UpdateUserRoleDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/users/{userID}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-User"
                ],
                "summary": "更新用戶狀態",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "狀態資訊",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_dto.UpdateUserStatusDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "interchange_internal_core.Role": {
            "type": "string",
            "enum": [
                "admin",
                "editor",
                "user",
                "readonly",
                "banned"
            ],
            "x-enum-comments": {
                "RoleAdmin": "管理員：可編輯所有人",
                "RoleBanned": "被禁用，無法登入或操作",
                "RoleEditor": "可編輯內容，但不能做高權限操作",
                "RoleReadOnly": "只能查詢，不能改資料",
                "RoleUser": "一般使用者"
            },
            "x-enum-varnames": [
                "RoleAdmin",
                "RoleEditor",
                "RoleUser",
                "RoleReadOnly",
                "RoleBanned"
            ]
        },
        "interchange_internal_core.Status": {
            "type": "string",
            "enum": [
                "active",
                "blocked",
                "suspended",
                "expired",
                "revoked",
                "maintenance",
                "pending",
                "deleted"
            ],
            "x-enum-comments": {
                "StatusActive": "正常可用",
                "StatusBlocked": "被封鎖（例如濫用）",
                "StatusDeleted": "已刪除（軟刪除）",
                "StatusExpired": "已過期",
                "StatusMaintenance": "系統維護中（暫時停用）",
                "StatusPending": "尚未啟用（等待審核/激活）",
                "StatusRevoked": "被手動撤銷",
                "StatusSuspended": "暫停（違規調查中）"
            },
            "x-enum-varnames": [
                "StatusActive",
                "StatusBlocked",
                "StatusSuspended",
                "StatusExpired",
                "StatusRevoked",
                "StatusMaintenance",
                "StatusPending",
                "StatusDeleted"
            ]
        },
...
        "interchange_internal_dto.CreateUserDto": {
            "type": "object",
            "required": [
                "displayName",
                "role",
                "status"
            ],
            "properties": {
                "displayName": {
                    "description": "顯示名稱",
                    "type": "string"
                },
                "email": {
                    "description": "信箱可選且格式驗證",
                    "type": "string"
                },
                "externalID": {
                    "description": "外部平台 ID（如 Google/Lark）",
                    "type": "string"
                },
                "lang": {
                    "description": "語系（如 zh-TW, en）",
                    "type": "string"
                },
                "role": {
                    "description": "角色",
                    "allOf": [
                        {
                            "$ref": "#/definitions/interchange_internal_core.Role"
                        }
                    ]
                },
                "status": {
                    "description": "狀態",
                    "allOf": [
                        {
                            "$ref": "#/definitions/interchange_internal_core.Status"
                        }
                    ]
                }
            }
        },
...
        "interchange_internal_dto.UpdateUserDto": {
            "type": "object",
            "properties": {
                "displayName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "externalID": {
                    "type": "string"
                },
                "lang": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/interchange_internal_core.Role"
                },
                "status": {
                    "$ref": "#/definitions/interchange_internal_core.Status"
                }
            }
        },
        "interchange_internal_dto.UpdateUserRoleDto": {
            "type": "object",
            "required": [
                "role"
            ],
            "properties": {
                "role": {
                    "$ref": "#/definitions/interchange_internal_core.Role"
                }
            }
        },
        "interchange_internal_dto.UpdateUserStatusDto": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "$ref": "#/definitions/interchange_internal_core.Status"
                }
            }
        },
...
        "interchange_internal_dto.UserResponseDto": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "externalID": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lang": {
                    "type": "string"
                },
                "last_seen": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/interchange_internal_core.Role"
                },
                "status": {
                    "$ref": "#/definitions/interchange_internal_core.Status"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "請在欄位輸入 \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}